Add files via upload
Introduction:
In this tutorial, we delve into the realm of machine learning by implementing a Random Forest Classifier using Python. Our focus lies on utilizing the Random Forest algorithm to analyze a bank dataset, which encompasses various features related to customers' attributes and banking activities. Through this tutorial, we aim to provide a comprehensive understanding of Random Forests and their application in predictive modeling.